requirements.txt.txt
streamlit
pandas
numpy
yfinance
statsmodels
scikit-learn
matplotlib
Pillow
